Fixed errors reported by gcc4.9

+++ avroq/filter/equality_expression.h
@@ -58,7 +58,7 @@ namespace detail {
         OP op;
         string_operator::ops_t strop;
         bool is_array_element = false;
-        int array_index = 0;
+        size_t array_index = 0;
         std::vector<bool> array_states;
